Discovered problems with variable assignment
Because OpenScad is a declarative programming language, I can't set the
variables during my nested if and else if statements.

I'm thinking of doing something funky like creating a new function that
inputs the values and then generates the supports and the other square.

These next iterations are going to be ugly, but atleast they have a shot
at working.